rename: layouts -> layout 변경
